Move commands added in 6.2 to `lib/command`
See #6421
