+++ wcfsetup/install/files/lib/acp/action/ToggleArticleI18nAction.class.php
@@ -9,8 +9,8 @@ use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Server\RequestHandlerInterface;
 use wcf\data\article\Article;
 use wcf\http\Helper;
-use wcf\system\article\command\DisableI18n;
-use wcf\system\article\command\EnableI18n;
+use wcf\command\article\DisableI18n;
+use wcf\command\article\EnableI18n;
 use wcf\system\exception\IllegalLinkException;
 use wcf\system\form\builder\field\RadioButtonFormField;
 use wcf\system\form\builder\LanguageItemFormNode;
