+++ wcfsetup/install/files/lib/command/file/ReplaceWithWebpVariant.class.php
@@ -1,8 +1,9 @@
 <?php
 
-namespace wcf\system\file\command;
+namespace wcf\command\file;
 
 use wcf\data\file\File;
+use wcf\system\file\command\ReplaceFileSource;
 
 /**
  * Converts an image to its WebP variant by replacing it with it.
