+++ wcfsetup/install/files/lib/command/user/SetAvatar.class.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace wcf\system\user\command;
+namespace wcf\command\user;
 
 use wcf\data\file\File;
 use wcf\data\file\FileAction;
