+++ wcfsetup/install/files/lib/system/file/processor/FileProcessor.class.php
@@ -14,7 +14,7 @@ use wcf\system\database\util\PreparedStatementConditionBuilder;
 use wcf\system\event\EventHandler;
 use wcf\system\exception\SystemException;
 use wcf\system\file\command\ReplaceFileSource;
-use wcf\system\file\command\ReplaceWithWebpVariant;
+use wcf\command\file\ReplaceWithWebpVariant;
 use wcf\system\file\processor\exception\DamagedImage;
 use wcf\system\image\adapter\exception\ImageNotProcessable;
 use wcf\system\image\adapter\exception\ImageNotReadable;
