+++ wcfsetup/install/files/lib/system/file/processor/UserAvatarFileProcessor.class.php
@@ -8,7 +8,7 @@ use wcf\data\user\UserProfile;
 use wcf\system\cache\runtime\UserProfileRuntimeCache;
 use wcf\system\database\util\PreparedStatementConditionBuilder;
 use wcf\system\exception\UserInputException;
-use wcf\system\user\command\SetAvatar;
+use wcf\command\user\SetAvatar;
 use wcf\system\WCF;
 use wcf\util\FileUtil;
